ecommerce-backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.1
aiosqlite==0.19.0
aiomysql==0.2.0
# aioodbc==0.3.3   # اقفله مؤقتًا
cryptography==42.0.2
greenlet==3.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Caching
redis==5.0.1

# Environment
python-dotenv==1.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Logging
loguru==0.7.2